pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-to-text"
version = "0.1.0"
description = "Convert iPhone audio recordings to text using OpenAI Whisper"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Speech-to-Text Contributors"}
]
keywords = ["speech-to-text", "whisper", "transcription", "audio", "korean", "iphone"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "openai-whisper>=20231117",
    "pydub>=0.25.1",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "types-click>=7.1.8",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/speech-to-text"
Documentation = "https://github.com/yourusername/speech-to-text#readme"
Repository = "https://github.com/yourusername/speech-to-text.git"
"Bug Tracker" = "https://github.com/yourusername/speech-to-text/issues"

[project.scripts]
speech-to-text = "speech_to_text.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
speech_to_text = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"